refactor files in proper package
