+++ lib/core/beamer/beamer_config.dart
